Se crearon repositorio, service e implementos para prodcuto, categoria e inventario

+++ src/main/java/hn/unah/lenguajes1900/delivery/delivery/entities/Categoria.java
@@ -15,7 +15,7 @@ public class Categoria {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idcategoria")
-    private Float idcategoria;
+    private Long idcategoria;
 
     private String nombre;
 }
